Policy: split tests into immediate_test.go and redrive_tes.go; keep lints green
